Tests/LibWeb/Text/expected/css/stylesheet-disabled-toggle-invalidation.txt

before disable: rgb(255, 0, 0)
after disable: rgb(0, 0, 0)
after re-enable: rgb(255, 0, 0)
